requirements.txt
gradio-client
Pillow
numpy
pytest